// src/src/route/router-event.enum.ts
export enum RouterEvent {
    NONE = 'NONE',
    START_CHANGE = 'START_CHANGE',
    FINISHED_CHANGE = 'FINISHED_CHANGE',
    CHANGE_ERROR = 'CHANGE_ERROR',
    STARTED_LOADING = 'STARTED_LOADING',
    FINISHED_LOADING = 'FINISHED_LOADING',
    COMPLETED_LOADING = 'COMPLETED_LOADING',
    ERROR_LOADING = 'ERROR_LOADING',
    STARTED_RERENDER = 'STARTED_RERENDER',
    FINISHED_RERENDER = 'FINISHED_RERENDER',
    REJECTED = 'REJECTED',
}